requirements.txt
python-dotenv>=1.0.1
pandas>=2.2.2
openpyxl>=3.1.5
langgraph>=0.2.34
langchain-core>=0.3.5
openai>=1.50.0
google-generativeai>=0.8.3
requests>=2.32.3
PyMuPDF>=1.24.10
GitPython>=3.1.43
nbformat>=5.10.4
pytest>=8.3.3
streamlit>=1.39.0
